# openapi/pjokk.yaml
openapi: 3.0.3
info:
  title: Pjokk API
  description: >-
    Pjokk is a self-hosted baby tracker for families. This spec is the single
    source of truth for the Go server's HTTP surface: it drives runtime
    request validation (kin-openapi nethttp-middleware), the generated
    strict-server routes and types (oapi-codegen), and the interactive docs
    served at /api/docs (Scalar, reading /api/openapi.json — this file,
    converted to JSON at runtime).
  version: "1.0.0"
servers:
  - url: /
paths:
...
  /api/sleep:
    get:
      operationId: listSleeps
      summary: Sleep logs in the caller's active family, newest first (by startTime).
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/babyIdQuery"
        - $ref: "#/components/parameters/limitQuery"
      responses:
        "200":
          description: Sleep logs, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepLog"
    post:
      operationId: createSleep
      summary: >-
        Create a sleep log. Omit endTime to start an active session; a baby
        can only have one active session at a time — enforced by a partial
        unique index (see internal/api/sleep.go), not just the pre-check
        this endpoint also does.
      tags: [sleep]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSleep"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepLog"
        "403":
          description: >-
            `caretakerId` names someone who is not a member of the caller's
            family (NOT_MEMBER).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Unknown baby.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Baby already has an active session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/sleep/active:
    get:
      operationId: getActiveSleep
      summary: The active sleep session for a baby, or null.
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/babyIdQuery"
      responses:
        "200":
          description: The active sleep session, or null.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SleepLog"
                nullable: true
  /api/sleep/{id}/wake:
    post:
      operationId: wakeSleep
      summary: End the active session for a sleep log. endTime defaults to now.
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Wake"
      responses:
        "200":
          description: Session ended.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepLog"
        "404":
          description: No such active session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/sleep/{id}:
    patch:
      operationId: updateSleep
      summary: >-
        Partial update. `endTime` sent as `null` CLEARS it, reopening the
        session — subject to the same one-active-session-per-baby
        constraint as create, so this can also 409. `location`/`notes` may
        also be sent as `null` to clear them; `startTime` is not nullable —
        only settable or omitted. See internal/api/feeds.go for the
        omitted-vs-null presence-detection pattern this endpoint needs.
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSleep"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepLog"
        "403":
          description: >-
            `caretakerId` names someone who is not a member of the caller's
            family (NOT_MEMBER).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No sleep log with this id in the caller's family.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >-
            Clearing endTime would reopen a session while another is
            already active for this baby.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSleep
      summary: Delete a sleep log.
      tags: [sleep]
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "404":
          description: No sleep log with this id in the caller's family.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    babyIdQuery:
      name: babyId
      in: query
      description: Restrict the result to one baby in the caller's family.
      required: false
      schema:
        type: string
    limitQuery:
      name: limit
      in: query
      description: Maximum number of rows to return.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
...
    idPath:
      name: id
      in: path
      description: Resource id.
      required: true
      schema:
        type: string
...
  schemas:
    Ok:
      type: object
      description: The standard `{ok:true}` success body for actions with no other payload.
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    SleepLog:
      type: object
      required:
        - id
        - babyId
        - caretakerId
        - caretakerName
        - loggedById
        - loggedByName
        - notes
        - startTime
        - endTime
        - location
        - type
      properties:
        id:
          type: string
        babyId:
          type: string
        caretakerId:
          type: string
        caretakerName:
          type: string
        loggedById:
          type: string
          description: Who saved the row; set by the server, never by a client.
        loggedByName:
          type: string
        notes:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
          description: null while the session is active.
        location:
          type: string
          nullable: true
        type:
          type: string
          enum: [nap, night]
          nullable: true
          description: >-
            Nap or night sleep. The client defaults it from the device's
            night-mode schedule; the server never guesses (it has no
            timezone), so an omitted type stays null. (Named `type` like
            the feed and diaper enums; `kind` is the timeline's
            discriminator and must stay free.)
    CreateSleep:
      type: object
      required: [babyId, startTime]
      properties:
        caretakerId:
          type: string
          description: >-
            Who did the care (as opposed to who is saving the row): a member of the caller's family, 403 NOT_MEMBER otherwise. Defaults to the caller.
        babyId:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          description: Omit to start an active sleep session.
        location:
          type: string
          maxLength: 100
        type:
          type: string
          enum: [nap, night]
        notes:
          type: string
          maxLength: 1000
    UpdateSleep:
      type: object
      description: >-
        Every field is optional; an empty object is a no-op. `endTime`,
        `location` and `notes` may also be sent as `null` to CLEAR that
        column — clearing `endTime` reopens the session and can 409 if
        another session for the same baby is already active (see
        internal/api/sleep.go). `startTime` is not nullable — only settable
        or omitted.
      properties:
        caretakerId:
          type: string
          description: >-
            Who did the care: a member of the caller's family (403 NOT_MEMBER otherwise). Not nullable — omit it to leave the caretaker unchanged.
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
          maxLength: 100
          nullable: true
        type:
          type: string
          enum: [nap, night]
          nullable: true
        notes:
          type: string
          maxLength: 1000
          nullable: true
    Wake:
      type: object
      description: Defaults endTime to now on the server when omitted.
      properties:
        endTime:
          type: string
          format: date-time
...
    Error:
      type: object
      description: >-
        The error envelope used by every non-2xx JSON response in the API
        (CLAUDE.md: "{error: string, code: string}" everywhere).
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        issues:
          type: array
          description: >-
            Optional per-field validation issues. The frontend only reads
            error/code today; issues exist for future, more specific UI.
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
              message:
                type: string